Split Monitor Controller and slider comp
